Move to receiver/modules/compose/ecs/utils
